test: move to the `string` and `number`.
