+++ src/string/test/string-state.spec.ts
@@ -1,4 +1,4 @@
-import { StringState } from "../lib";
+import { StringState } from "../../lib";
 
 // Extend the `StringState` to create a custom class
 class MyStringState<Type extends string = string> extends StringState<Type> {
@@ -12,6 +12,7 @@ class MyStringState<Type extends string = string> extends StringState<Type> {
   }
 }
 
+
 console.group(`StringState`);
 
 // Instantiate your custom StringState class
@@ -22,18 +23,18 @@ console.log(myState.state);  // Output: Hello World
 
 // Change the state and store the history
 myState.set("Goodbye World");
-console.log(myState.state);  // Output: Goodbye World
+console.log(``, myState.state);  // Output: Goodbye World
 
 // Undo the change to revert to the previous state
-myState.undo();
-console.log(myState.state);  // Output: Hello World
+myState.history?.undo();
+console.log(`after undo`, myState.state);  // Output: Hello World
 
 // Redo the change to restore the state
-myState.redo();
-console.log(myState.state);  // Output: Goodbye World
+myState.history?.redo();
+console.log(``, myState.state);  // Output: Goodbye World
 
 // Set a new state and see the history update
 myState.set("New State");
-console.log(myState.state);  // Output: New State
+console.log(``, myState.state);  // Output: New State
 
 console.groupEnd();
